# Conflicts:
#   src/main/java/edu/ntnu/idi/stud/team10/sparesti/controller/BankController.java
#   src/main/java/edu/ntnu/idi/stud/team10/sparesti/dto/BudgetDto.java
#   src/main/java/edu/ntnu/idi/stud/team10/sparesti/dto/UserDto.java

+++ src/main/java/edu/ntnu/idi/stud/team10/sparesti/dto/TransactionDto.java
@@ -16,6 +16,6 @@ public class TransactionDto {
   private double amount;
   private String description;
   private String category;
-  private int accountNr;
+  private Long accountNr;
   private LocalDate date;
 }
